also check if guild still exists

+++ source/raid_cog.py
@@ -603,9 +603,10 @@ class RaidCog(commands.Cog):
         logger.info(message)
         tag, guild_id = select_one(self.conn, 'Raids', ['tag', 'guild_id'], ['raid_id'], [raid_id])
         guild = self.bot.get_guild(guild_id)
-        role = discord.utils.get(guild.roles, name=tag)
-        if role:
-            await role.delete()
+        if guild:
+            role = discord.utils.get(guild.roles, name=tag)
+            if role:
+                await role.delete()
         delete(self.conn, 'Raids', ['raid_id'], [raid_id])
         delete(self.conn, 'Players', ['raid_id'], [raid_id])
         delete(self.conn, 'Assignment', ['raid_id'], [raid_id])
